+++ CHANGES.rst
@@ -20,6 +20,7 @@ development
 - [mw] Improve report generation with ``unoconv``
 - [license] Officially upgrade to EUPL 1.2
 - [ui] Improve full text display for IFI Claims
+- [mw] Fix full text acquisition from Espacenet
 
 2018-03-17 0.150.0
 ==================